add base tests for bankpage.ts
